feat: add kazeterm-event-system crate for centralized event handling
- Introduced a new crate `kazeterm-event-system` to manage application events.
- Defined `AppEvent` enum for various terminal actions.
- Implemented `EventBus` for event dispatching and subscription management.
- Added event sources for reading from stdin and Unix domain sockets.
- Integrated event handling into the Kazeterm application, allowing for external event input.
- Updated `Cargo.toml` files to include the new crate and its dependencies.
- Refactored existing event handling code to utilize the new event system.
