+++ crates/kazeterm-event-system/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "kazeterm-event-system"
+version = "0.1.0"
+publish.workspace = true
+edition.workspace = true
+
+[dependencies]
+anyhow.workspace = true
+gpui.workspace = true
+serde.workspace = true
+serde_json = "1"
+smol.workspace = true
+tracing.workspace = true
+
+[target.'cfg(target_os = "windows")'.dependencies]
+uds_windows = "1"
+
+[lints]
+workspace = true
