+++ crates/kazeterm-event-system/src/json_event.rs
@@ -2,22 +2,22 @@ use std::path::PathBuf;
 
 use serde::Deserialize;
 
-use super::AppEvent;
+use crate::AppEvent;
 
-/// Configuration for the event source
-#[derive(Debug, Clone, Default)]
+/// Configuration for the external event source.
+#[derive(Debug, Clone, Default, PartialEq, Eq)]
 pub enum EventSourceConfig {
-  /// No external event source (events can still be sent programmatically)
+  /// No external event source (events can still be sent programmatically).
   #[default]
   None,
-  /// Read events from stdin (JSON, one per line)
+  /// Read events from stdin (JSON, one per line).
   Stdio,
-  /// Read events from a Unix domain socket (all platforms)
+  /// Read events from a Unix domain socket (all platforms).
   Socket { path: PathBuf },
 }
 
-/// JSON representation of an event for external input
-#[derive(Debug, Deserialize)]
+/// JSON representation of an event for external input.
+#[derive(Debug, Clone, Deserialize, PartialEq, Eq)]
 #[serde(tag = "event")]
 pub enum JsonEvent {
   NewTerminalWithDefaultProfile,
@@ -39,12 +39,20 @@ pub enum JsonEvent {
   CloseActivePane,
   FocusNextPane,
   FocusPreviousPane,
+  FocusPaneUp,
+  FocusPaneDown,
+  FocusPaneLeft,
+  FocusPaneRight,
   SwapSplitPanes,
   ToggleSearch,
+  ToggleFullscreen,
   ToggleTabBar,
   ShowAboutDialog,
+  ShowImportAlacrittyDialog,
   ReloadConfig,
   FocusActiveTerminal,
+  NewWindow,
+  Quit,
   SendTextToTerminal {
     text: String,
   },
@@ -75,12 +83,20 @@ impl From<JsonEvent> for AppEvent {
       JsonEvent::CloseActivePane => AppEvent::CloseActivePane,
       JsonEvent::FocusNextPane => AppEvent::FocusNextPane,
       JsonEvent::FocusPreviousPane => AppEvent::FocusPreviousPane,
+      JsonEvent::FocusPaneUp => AppEvent::FocusPaneUp,
+      JsonEvent::FocusPaneDown => AppEvent::FocusPaneDown,
+      JsonEvent::FocusPaneLeft => AppEvent::FocusPaneLeft,
+      JsonEvent::FocusPaneRight => AppEvent::FocusPaneRight,
       JsonEvent::SwapSplitPanes => AppEvent::SwapSplitPanes,
       JsonEvent::ToggleSearch => AppEvent::ToggleSearch,
+      JsonEvent::ToggleFullscreen => AppEvent::ToggleFullscreen,
       JsonEvent::ToggleTabBar => AppEvent::ToggleTabBar,
       JsonEvent::ShowAboutDialog => AppEvent::ShowAboutDialog,
+      JsonEvent::ShowImportAlacrittyDialog => AppEvent::ShowImportAlacrittyDialog,
       JsonEvent::ReloadConfig => AppEvent::ReloadConfig,
       JsonEvent::FocusActiveTerminal => AppEvent::FocusActiveTerminal,
+      JsonEvent::NewWindow => AppEvent::NewWindow,
+      JsonEvent::Quit => AppEvent::Quit,
       JsonEvent::SendTextToTerminal { text } => AppEvent::SendTextToTerminal { text },
       JsonEvent::Custom { name, data } => AppEvent::Custom { name, data },
     }
